se termino con la configuracion de hibernate, reservas-core y usuarioBO.

+++ reservas-core/src/main/java/com/reservas/dao/UsuarioDAO.java
@@ -2,6 +2,6 @@ package com.reservas.dao;
 
 import com.reservas.model.UsuarioBO;
 
-public interface UsuarioDAO extends BaseDAO<Integer, UsuarioBO> {
+public interface UsuarioDAO extends BaseDAO<Long, UsuarioBO> {
 
 }
