+++ reservas-core/src/main/java/com/reservas/service/UsuarioService.java
@@ -2,6 +2,6 @@ package com.reservas.service;
 
 import com.reservas.model.UsuarioBO;
 
-public interface UsuarioService extends BaseService<Integer, UsuarioBO> {
+public interface UsuarioService extends BaseService<Long, UsuarioBO> {
 
 }
